check marks appear when individual sets saved
It doesn't work as expected, all the time! Because we can't save individual sets, but instead try saving all sets in a form at the same time, there are some strange quirks. For example, if there's an error saving the form, we don't know if it's a problem with the individual set, or with all the sets as a whole. The alert box that pops up when a workout is saved more accurately represents the internal state of our app. One way to fix this would be to change the model so that a "set" is an object, or a table in the database. Each workout could contain multiple links to a different sets. This would enable us to create set objects, and save them to a workout, and do so individually as users fill out sets line by line, as opposed to trying to save the whole workout each time.