Check-marks on autosave. Cleaned up all other code. Fixes #13, #14
The workouts now save on key-up, meaning it's very unlikely that there is ever a point at which they are unsaved. This is good.

The check marks only appear on key-up if the whole workout form is valid. If the workout form is invalid, the website displays a spinning icon (to show that it's trying to submit), and then displays nothing. It might be a good idea to show a red "x" to indicate that it is unsaved, but this is a problem because we have no way of saving individual sets, and it would be misleading to say that the set is "bad" if the real culprit is another set that has invalid input.

I fixed the expanding/collapsing workouts functions. It used to be that there was an event handler attached to the click event for a workout title that expanded the workout form, then, it would remove that handler, and attach other one that would contract and save the workout form on the next click. This caused all sorts of crazy problems (if the event was interupted and never attached the new event, then the workout form would never collapse). In order to fix, I made the onclick event handler for the workout_name element test to see if it was expanded or contracted, and then do one of two things: 1) expand the form, 2) contract and submit the form. This works much better.