Label to indicate which workouts have been added, instead of divider
It used to just have a thick line between workouts that have been added and workouts that haven't been added. Julian was confused by this. I set him straight by adding labels in the list. "Today's workouts" and "More workouts" to clearly indicate that some of the workouts have been added to "today," and some of the workouts have not.